Adding hooks to fix visual bugs.
